Não tou a entender nada agora

+++ src/server/exec.c
@@ -1,7 +1,7 @@
 #include "exec.h"
 
 // Função que constroi uma struct de execução
-Exec buildExec(int pid, int tempo, char *nome){
+Exec buildExec(int pid, long tempo, char *nome){
     Exec exe = malloc(sizeof(struct execucao));
 
     exe->pid = pid;
@@ -12,5 +12,5 @@ Exec buildExec(int pid, int tempo, char *nome){
 }
 
 void printaExec(Exec exe){
-    printf("PID: %d\nNome: %s\nTempo: %d\n", exe->pid, exe->nome, exe->tempo);
+    printf("PID: %d\nNome: %s\nTempo: %lli\n", exe->pid, exe->nome, exe->tempo);
 }
